+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Secure default options for native PHP sessions.
 - Package documentation.
 
+### Breaking
+
+- Remembered URI redirects now only return safe local paths.
+
 ## [0.1.0](https://github.com/duoncode/session/releases/tag/0.1.0) (2026-01-31)
 
 Initial release.